requirements.txt
python-dotenv

uvicorn~=0.22.0
fastapi~=0.95.1
pydantic~=1.10.7
SQLAlchemy~=2.0.18
starlette~=0.26.1